// src/app/pages/adicionar-formulario/enums/TypeQuestEnum.ts
export enum TypeQuestEnum {
  TEXTO = "TEXTO",                // texto curto
  PARAGRAFO = "PARAGRAFO",        // texto longo
  NUMERO = "NUMERO",              // número
  UNICA = "UNICA",                // múltipla escolha (radio)
  MULTIPLA = "MULTIPLA",          // múltipla escolha (checkbox)
  CHECKBOX = "CHECKBOX",          // mesmo que MULTIPLA
  DATA = "DATA",                  // só data
  DATAHORA = "DATAHORA",          // data e hora
  ESCALA = "ESCALA",              // escala / nota
  VERDADEIRO_FALSO = "VERDADEIRO_FALSO", // verdadeiro/falso
  UPLOAD = "UPLOAD",              // upload de arquivos
}